Fix cvxpy failure in nightly testing (#656)
When building the optimization problem from Papilo, the problem category was set to `LP` regardless of the type of problem. This caused the size of `presolve_data.var_flags` to be zero, leading to a illegal memory access in `set_implied_integers`. 

Closes #559.

Authors:
  - Nicolas L. Guidotti (https://github.com/nguidotti)

Approvers:
  - Hugo Linsenmaier (https://github.com/hlinsen)

URL: https://github.com/NVIDIA/cuopt/pull/656